decoupled Queue and LL classes
