+++ Scheduler.java
@@ -1,7 +1,5 @@
 
-public class Scheduler extends LinkedList{
-	
-	
+public class Scheduler{
 	
 	public static LinkedList sortOneList(LinkedList unsorted){
 		
@@ -12,25 +10,20 @@ public class Scheduler extends LinkedList{
 			sorted.insert(walker.getPcb(), sorted);
 			walker = walker.getNext();
 		}
-		
 		return sorted;
 	}
 	
 	
 	public static LinkedList sortTwoList(LinkedList unsorted1, LinkedList unsorted2){
-		
 		LinkedList sorted = sortOneList(unsorted1);
-		
 		Node walker2 = unsorted2.head;
 		
 		while(walker2!=null){
 			sorted.insert(walker2.getPcb(), sorted);
 			walker2 = walker2.getNext();
 		}
-		
 		return sorted;
 	}
-	
 	
 	
 }
